% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submission.R
\docType{data}
\name{submission_server}
\alias{submission_server}
\alias{submission_ui}
\title{Tutorial submission functions}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
submission_server(input, output)

submission_ui
}
\description{
Based on Professor Matt Blackwell's and Professor Colin Rundel's work
https://github.com/mattblackwell/qsslearnr and
https://github.com/rundel/learnrhash

From their notes:
When including these functions in a learnr Rmd document it is necessary that the
server function, `submission_server()`, be included in an R chunk where
`context="server"`. Conversely, any of the ui functions, `*_ui()`, must *not*
be included in an R chunk with a `context`.
}
\keyword{datasets}
